+++ environment/src/test/java/jetbrains/exodus/log/LogTests.java
@@ -42,7 +42,7 @@ public class LogTests extends LogTestsBase {
         super.setUp();
         LoggableFactory.registerLoggable(DUMMY_LOGGABLE.getType(), new LoggableFactory() {
             @Override
-            protected RandomAccessLoggable create(long address, int length, @NotNull ByteIterableWithAddress data, int dataLength, long structureId) {
+            protected RandomAccessLoggable create(long address, int length, @NotNull ByteIterableWithAddress data, int dataLength, int structureId) {
                 return new RandomAccessLoggableImpl(address, DUMMY_LOGGABLE.getType(), length, ByteIterableWithAddress.EMPTY, dataLength, structureId);
             }
         });
@@ -187,7 +187,7 @@ public class LogTests extends LogTestsBase {
             public void run() {
                 final LoggableFactory factory = new LoggableFactory() {
                     @Override
-                    protected RandomAccessLoggable create(long address, int length, @NotNull ByteIterableWithAddress data, int dataLength, long structureId) {
+                    protected RandomAccessLoggable create(long address, int length, @NotNull ByteIterableWithAddress data, int dataLength, int structureId) {
                         throw new UnsupportedOperationException("Shouldn't be there");
                     }
                 };
